refactor: isolate models report CSV rendering
