+++ src/models-report.ts
@@ -8,7 +8,9 @@ import { getProvider } from './providers/index.js'
 import { CATEGORY_LABELS, type ProjectSummary, type TaskCategory } from './types.js'
 import { combineReasoningSemantics, reasoningSemanticsForProviders, reasoningTokenTotals, type ReasoningTokenSemantics } from './token-semantics.js'
 import type { AggregateOptions, ModelReportRow } from './models-report-types.js'
+import { renderCsv } from './models-report-csv.js'
 export type { AggregateOptions, ModelReportRow } from './models-report-types.js'
+export { renderCsv }
 
 type Bucket = {
   provider: string
@@ -585,13 +587,6 @@ export function renderJson(rows: ModelReportRow[]): string {
   )
 }
 
-function csvEscape(value: string | undefined | null): string {
-  if (value === undefined || value === null) return ''
-  if (value.includes(',') || value.includes('"') || value.includes('\n')) {
-    return `"${value.replace(/"/g, '""')}"`
-  }
-  return value
-}
 
 function mdEscape(value: string): string {
   // Pipes break GitHub-flavored markdown tables; escape them.
@@ -669,73 +664,5 @@ export function renderMarkdown(rows: ModelReportRow[], opts: { byTask?: boolean;
     lines.push(`| ${totalCells.join(' | ')} |`)
   }
 
-  return lines.join('\n')
-}
-
-export function renderCsv(rows: ModelReportRow[], opts: { byTask?: boolean; byAgent?: boolean } = {}): string {
-  const byTask = opts.byTask ?? false
-  const byAgent = opts.byAgent ?? false
-  // CSV intentionally repeats provider/model on every row so downstream
-  // consumers can sort/filter without first reconstructing the grouping.
-  const header = byAgent
-    ? ['provider', 'model', 'agent', 'input_tokens', 'output_tokens', 'reasoning_tokens', 'additive_reasoning_tokens', 'cache_write_tokens', 'cache_read_tokens', 'total_tokens', 'calls', 'cost_usd', 'savings_usd', 'savings_baseline_model']
-    : byTask
-    ? ['provider', 'model', 'task', 'input_tokens', 'output_tokens', 'reasoning_tokens', 'additive_reasoning_tokens', 'cache_write_tokens', 'cache_read_tokens', 'total_tokens', 'calls', 'cost_usd', 'savings_usd', 'savings_baseline_model']
-    : ['provider', 'model', 'top_task', 'top_task_share', 'input_tokens', 'output_tokens', 'reasoning_tokens', 'additive_reasoning_tokens', 'cache_write_tokens', 'cache_read_tokens', 'total_tokens', 'calls', 'cost_usd', 'savings_usd', 'savings_baseline_model']
-  const lines: string[] = [header.join(',')]
-  for (const r of rows) {
-    const cells = byAgent
-      ? [
-          csvEscape(r.providerDisplayName),
-          csvEscape(r.modelDisplayName),
-          csvEscape(r.agentType ?? ''),
-          String(r.inputTokens),
-          String(r.outputTokens),
-          String(r.reasoningTokens ?? 0),
-          String(r.additiveReasoningTokens ?? 0),
-          String(r.cacheWriteTokens),
-          String(r.cacheReadTokens),
-          String(r.totalTokens),
-          String(r.calls),
-          r.costUSD.toFixed(6),
-          (r.savingsUSD ?? 0).toFixed(6),
-          csvEscape(r.savingsBaselineModel),
-        ]
-      : byTask
-      ? [
-          csvEscape(r.providerDisplayName),
-          csvEscape(r.modelDisplayName),
-          r.category ? categoryLabel(r.category) : '',
-          String(r.inputTokens),
-          String(r.outputTokens),
-           String(r.reasoningTokens ?? 0),
-          String(r.additiveReasoningTokens ?? 0),
-          String(r.cacheWriteTokens),
-          String(r.cacheReadTokens),
-          String(r.totalTokens),
-          String(r.calls),
-          r.costUSD.toFixed(6),
-          (r.savingsUSD ?? 0).toFixed(6),
-          csvEscape(r.savingsBaselineModel),
-        ]
-      : [
-          csvEscape(r.providerDisplayName),
-          csvEscape(r.modelDisplayName),
-          r.topCategory ? categoryLabel(r.topCategory) : '',
-          r.topCategoryShare !== undefined ? r.topCategoryShare.toFixed(4) : '',
-          String(r.inputTokens),
-          String(r.outputTokens),
-          String(r.reasoningTokens ?? 0),
-          String(r.additiveReasoningTokens ?? 0),
-          String(r.cacheWriteTokens),
-          String(r.cacheReadTokens),
-          String(r.totalTokens),
-          String(r.calls),
-          r.costUSD.toFixed(6),
-          (r.savingsUSD ?? 0).toFixed(6),
-          csvEscape(r.savingsBaselineModel),
-        ]
-    lines.push(cells.join(','))
-  }
   return lines.join('\n')
 }
